EGS-669: Implement IP ranges
This commit implements first stage for ranges support, for now just for IP addresses, that allows to use ranges as part of the `in` operator. Example: `ip.addr in { 1.1.1.1 1.2.3.4..5.6.7.8 10.0.0.0/16 }` (where CIDR is just a syntax sugar for an extended range).

Also we decided that we no longer want `ip.addr == 1.2.3.4/8` to be valid but instead encourage using `ip.addr in { 1.2.3.4/8 }` to be consistent with other range types.

For now, these are implemented as simple `Vec`s of items or ranges that intentionally preserve original order of items, but we can change that later for performance.