Simplify Bytes type
Stop distinguishing between Raw and Str variants of the input.

This was needed mainly for Debug formatting, so now it will simply print both possible representation.

Another place that made a distinction was Regex enforcing unicode boundaries on text-like expressions, but as we're dealing with any text input as raw bytes, this is not necessary either and allows to keep implementation cleaner.