Switch to Criterion for benchmarking
This allows to run benchmarks on stable Rust, and also gives better statistical confidence in performance improvements (and nice graphs!).

Currently benchmarks are changed to only test field comparison operations, as these are much more expensive and so more interesting than `and`/`or`/`xor`/`not`, but we can always add the latter in future too.

These benchmarks intend to compare different combinations of field values and filters to cover different cases such as filter passing, filter not passing and IPv4 vs IPv6 mismatch.

These combinations run as individual benchmarks with dedicated warmup and measurement times, and so running all benchmarks is much slower as a result, but returned timings are much more precise, and Criterion allows to filter out and rerun only benchmarks you are interested in and so get precise performance improvements after changes.