Use binary search in RangeSet
Now that it's not exposed in AST, we can optimise it in whichever way we want.

After this change, range sets are going to be presorted during compilation and use binary search during execution.

This brings execution complexity to O(ln N) instead of O(N), which slows down "best" edge cases where IP range previously happened to be one of the first, but instead significantly (2x) speeds up average case with arbitrary lookup, especially on large sets.