Re-add TryFrom implementation to LhsValue
This was deleted when working on the Map & Array data types.
See e500dc1f5f80c38273d2436f9fe82cff7fef3c40 for more info.

Before this commit `TypeMismatchError` was expecting fully qualified
`Type` for the expected `member`. However when executing `Map::try_form`
we only know that we expect a Map but not a Map of which specific type
(same apply for Array). To overcome this issue, i thought about two
different solution:
* Implement custom `try_from(ty: Type)` method for Map and Array
  without implementing the TryFrom trait itself (because the prototype
  of try_from does not expect any argument)
* Modify the `TypeMismatchError` type to be able to express that
  lack of knowledge.

I chose the later:
* ExpectedMismatchType::Type(Type) corresponds to a fully qualified
  expected type
* ExpectedMismatchType::Array is used when expected an Array bot not
  knowing of which Type
* ExpectedMismatchType::Map is used when expected a Map bot not
  knowing of which Type