Make Filter implicitly Sync + Send
This is achieved by adding Sync + Send trait bounds to the Compiled*Expr
closure types.

This required a change to HeapSearcher to store a boxed pointer instead
of a raw pointer (which is !Sync). The unsafe code used to detach the
pointer from its lifetime changes, but the semantics are identical.