Engine: Regex: Avoid compiling and storing the same regex multiple times
Previously, the engine was compiling the same regex multiple times
during execution. This redundant compilation process led to unnecessary
performance overhead and increased resource consumption.

With this change, we store the compiled regex in a `HashSet`. If the
same regex is used more than once, the engine reuses the previously
compiled instance.

Signed-off-by: Jon Doron <jond@wiz.io>