Rework `Scheme` serialization
This change is a necessary rework of scheme serialization to allow
for additional field properties that will be introduced in a future
commit.

Before this change, a serialized scheme was a map of field name to field type.
After this change, a serialized scheme is a map of field name to
field definition object, which only contain its type for now but can
be extended with additional properties.